requirements.txt
asgiref==3.4.1
black==21.7b0
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.6.1
install==1.3.4
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0
zipp==3.5.0
